Moved the muntinlupaLocations, calculateDistance and calculateFare in LocationContext
